feat(ci): enforce the 800-line file limit and a ratcheting per-folder file limit
CLAUDE.md has required "max 800 LOC per file" all along, and scripts/check-file-length.sh
already implemented it — but it was wired into nothing. No workflow ran it, which is
how handlers.go reached 977 lines. A rule nothing enforces is a suggestion.

Adds a Structure Gates CI job running both limits.

**800 lines per file.** The script now excludes generated code (*/generated/*),
build output (extension/), and agent worktrees (.claude/). Those are not organized
by hand, so "split them" is meaningless advice and an unfixable violation only
teaches people to ignore the gate. That exclusion takes the list from 5 files to the
3 real ones, all fixed here along real seams:

- handlers.go 977 -> 604: the WebSocket transport (upgrade handshake, three-goroutine
  relay loop, shared deadline-bound frame writer) moves to ws.go. One cohesive
  concern, separable from session lifecycle and asset serving.
- sidepanel.ts 1020 -> 770: host-tab resolution and the two actions needing it move
  to panel/host-tab.ts (no panel state at all); the header/shell DOM construction
  moves to panel/shell.ts. shell.ts owns NO state — every element it builds is
  handed back through ShellDeps setters and every action is injected, with one local
  adapter keeping call sites unchanged. It is a sink rather than a plain return
  because the iframe setter fires from load/error handlers.
- tracked-hover-launcher.ts 801 -> 752: the shutter sound and capture flash move to
  hover/screenshot-feedback.ts. The caller was priming the AudioContext inline during
  the click, so the module exports primeShutterAudio() rather than leaking the
  variable back out — Chrome blocks audio created outside a user gesture.

**10 source files per folder, ratcheting.** A hard limit is unshippable today: 29
folders exceed it and cmd/browser-agent has 195 files. Every over-limit folder is
frozen at its current count in .folder-size-baseline.json; a folder may never grow
past its budget, and any folder NOT on the baseline gets the real 10-file limit. So
every PR can only improve things, and a folder that reaches 10 drops off the baseline
permanently and can never regress. A third CI step re-runs the generator and fails if
the committed baseline is stale, without which the ratchet would silently stop
tightening.

Verified the gate actually fails, since one that cannot is theater: a 196th file in
cmd/browser-agent fails on its baseline, and a new folder with 11 files fails on the
limit. It also caught a real regression immediately — internal/pty had reached 11
files (diag.go, added after the baseline was generated). upload.go moves to its own
internal/pty/upload package: it references no Session, Manager, Fanout or WriteBuffer,
so it shares nothing with the session primitives and had exactly one caller.

The two gates pull against each other by design — splitting a file for the LOC limit
adds a file to its folder. internal/terminal was at 9 of 10 and src/content/ui at 10,
which is why the TS extractions went into new hover/ and panel/ subfolders rather than
more siblings. That pressure is the point: it forces organization into sub-modules
instead of ever-wider flat directories.

eslint.config.js gains a commonjs block for scripts/**/*.cjs — the repo is
"type": "module", so a Node script needing require/__dirname must be .cjs, and it was
being linted as ESM (13 no-undef errors).

Gates: go vet clean; pty + pty/upload + terminal under -race; coverage 70.7% (gate 70%);
typecheck + compile-ts; test:ext 2979 passed / 0 failed; eslint clean; verify-llm passed;
both structure gates green; docs cross-ref updated (rule 9).
