+++ .folder-size-baseline.json
@@ -1,0 +1,35 @@
+{
+  "_comment": "Ratcheting baseline for scripts/check-folder-size.cjs. A folder may never exceed its recorded count; folders absent here are held to the 10-file limit. Counts may only go DOWN — run `make folder-baseline-update` after reducing one.",
+  "max_files": 10,
+  "folders": {
+    "cmd/browser-agent": 195,
+    "cmd/browser-agent/internal/bridge": 22,
+    "cmd/browser-agent/internal/toolconfigure": 14,
+    "cmd/browser-agent/internal/toolgenerate": 13,
+    "cmd/browser-agent/internal/toolinteract": 42,
+    "internal/analysis": 28,
+    "internal/capture": 67,
+    "internal/checkpoint": 11,
+    "internal/mcp": 12,
+    "internal/noise": 13,
+    "internal/queries": 15,
+    "internal/recording": 16,
+    "internal/schema": 17,
+    "internal/security": 43,
+    "internal/session": 20,
+    "internal/streaming": 11,
+    "internal/testgen": 11,
+    "internal/tools/analyze": 13,
+    "internal/tools/configure": 15,
+    "internal/tools/observe": 32,
+    "internal/upload": 19,
+    "internal/util": 16,
+    "scripts": 13,
+    "scripts/docs": 16,
+    "src/background": 35,
+    "src/background/dom": 11,
+    "src/lib": 35,
+    "src/popup": 14,
+    "src/types": 22
+  }
+}
